refactor: :truck: Rename shared_prefs to favorites_shared_prefs
